Moved InitialCreditsExpirationService logic into newly created service.
