--- api/src/main/java/org/pmiops/workbench/db/dao/InitialCreditsExpirationService.java
@@ -1,7 +1,0 @@
-package org.pmiops.workbench.db.dao;
-
-import java.util.List;
-
-public interface InitialCreditsExpirationService {
-  void checkCreditsExpirationForUserIDs(List<Long> userIdsList);
-}
