+++ api/src/main/java/org/pmiops/workbench/initialcredits/InitialCreditsExpirationService.java
@@ -1,9 +1,11 @@
 package org.pmiops.workbench.initialcredits;
 
 import java.sql.Timestamp;
+import java.util.List;
 import java.util.Optional;
 import org.pmiops.workbench.db.model.DbUser;
 
 public interface InitialCreditsExpirationService {
   Optional<Timestamp> getCreditsExpiration(DbUser user);
+  void checkCreditsExpirationForUserIDs(List<Long> userIdsList);
 }
